fixed OpenCL ICC Warnings
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2812 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/basis/linear/noboundary/operation/OperationMultipleEvalIterativeHybridX86SimdOCLLinear.cpp
@@ -125,8 +125,8 @@ namespace sg {
 #ifdef _OPENMP
         } else {
 #endif
-          int worksize = static_cast<int>((storageSize - gpu_partition) / (num_threads));
-          int myStart = static_cast<int>(gpu_partition + (tid - 1) * worksize);
+          size_t worksize = static_cast<size_t>((storageSize - gpu_partition) / (num_threads));
+          size_t myStart = static_cast<size_t>(gpu_partition + (tid - 1) * worksize);
           size_t myEnd = static_cast<size_t>(myStart + worksize);
 
           if (tid == num_threads - 1)
@@ -143,7 +143,7 @@ namespace sg {
           //      }
 #if defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t j = (size_t)myStart; j < myEnd; j++) {
+          for (size_t j = myStart; j < myEnd; j++) {
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
 
@@ -216,7 +216,7 @@ namespace sg {
 #endif
 #if defined(__SSE3__) && defined(__AVX__)
 
-          for (size_t j = (size_t)myStart; j < myEnd; j++) {
+          for (size_t j = myStart; j < myEnd; j++) {
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
 
@@ -291,7 +291,7 @@ namespace sg {
 #endif
 #if !defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t j = (size_t)myStart; j < myEnd; j++) {
+          for (size_t j = myStart; j < myEnd; j++) {
             ptrGlobalResult[j] = 0.0;
 
             for (size_t i = 0; i < source_size; i++) {
@@ -383,14 +383,14 @@ namespace sg {
             gpu_time = omp_get_wtime() - loc_start;
           }
         } else {
-          int worksize = static_cast<int>((result_size - gpu_partition) / (num_threads));
+          size_t worksize = static_cast<size_t>((result_size - gpu_partition) / (num_threads));
 #if defined(__SSE3__) && !defined(__AVX__)
           worksize = (worksize / 8) * 8;
 #endif
 #if defined(__SSE3__) && defined(__AVX__)
           worksize = (worksize / 16) * 16;
 #endif
-          int myStart = static_cast<int>(gpu_partition + (tid - 1) * worksize);
+          size_t myStart = static_cast<size_t>(gpu_partition + (tid - 1) * worksize);
           size_t myEnd = static_cast<size_t>(myStart + worksize);
 
           if (tid == num_threads - 1)
@@ -407,7 +407,7 @@ namespace sg {
           //      }
 #if defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t i = (size_t)myStart; i < myEnd; i += 8) {
+          for (size_t i = myStart; i < myEnd; i += 8) {
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
 
@@ -492,7 +492,7 @@ namespace sg {
 #endif
 #if defined(__SSE3__) && defined(__AVX__)
 
-          for (size_t i = (size_t)myStart; i < myEnd; i += 16) {
+          for (size_t i = myStart; i < myEnd; i += 16) {
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
 
@@ -577,7 +577,7 @@ namespace sg {
 #endif
 #if !defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t i = (size_t)myStart; i < myEnd; i++) {
+          for (size_t i = myStart; i < myEnd; i++) {
             for (size_t j = 0; j < storageSize; j++) {
               double curSupport = ptrAlpha[j];
